Reorganize code; add gas value asserts
